+++ examples/simple/settings/testing_mysql.py
@@ -1,0 +1,3 @@
+from .testing import *
+
+DATABASES = MYSQL_DATABASES