+++ examples/simple/settings/testing_postgresql.py
@@ -1,0 +1,3 @@
+from .testing import *
+
+DATABASES = POSTGRESQL_DATABASES