+++ examples/simple/settings/testing_sqlite.py
@@ -1,0 +1,3 @@
+from .testing import *
+
+DATABASES = SQLITE_DATABASES