+++ src/ormex/__init__.py
@@ -3,9 +3,9 @@ Django ORM extensions.
 """
 
 __title__ = 'django-ormex'
-__version__ = '0.1.1'
+__version__ = '0.2'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2017 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('default_app_config',)