+++ src/ormex/aggregations/group_concat/__init__.py
@@ -3,7 +3,7 @@ from nine.versions import DJANGO_LTE_1_9
 if DJANGO_LTE_1_9:
     from .django_lte_1_9 import *
 else:
-    from .django_gt_1_10 import *
+    from .django_gte_1_10 import *
 
 __title__ = 'ormex.aggregations.group_concat'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'