+++ src/ormex/aggregations/group_concat/django_lte_1_9.py
@@ -51,10 +51,26 @@ class PostgreSQL9GroupConcat(SQLAggregate):
 
     sql_function = 'string_agg'
 
+    def __init__(self, *expressions, **extra):
+        # For PostgreSQL separator is an obligatory
+        if 'separator' not in extra:
+            extra.update({'separator': ', '})
+        super(PostgreSQL9GroupConcat, self).__init__(*expressions, **extra)
+
     @property
     def sql_template(self):
-        # The ::text cast is a hardcoded hack to work with integer columns
-        return "%(function)s(%(field)s::text, '%(separator)s')"
+        # The ::text cast is a hardcoded hack to work with integer columns.
+        # Also, separator is obligatory
+        sort_results = self.extra.get('sort_results', False)
+
+        if sort_results:
+            return "%(function)s(" \
+                   "%(field)s::text, " \
+                   "'%(separator)s' " \
+                   "ORDER BY %(field)s" \
+                   ")"
+        else:
+            return "%(function)s(%(field)s::text, '%(separator)s')"
 
 
 class PostgreSQL8GroupConcat(SQLAggregate):
@@ -65,9 +81,23 @@ class PostgreSQL8GroupConcat(SQLAggregate):
 
     sql_function = 'array_to_string'
 
+    def __init__(self, *expressions, **extra):
+        # For PostgreSQL separator is an obligatory
+        if 'separator' not in extra:
+            extra.update({'separator': ', '})
+        super(PostgreSQL9GroupConcat, self).__init__(*expressions, **extra)
+
     @property
     def sql_template(self):
-        return "%(function)s(array_agg(%(field)s), '%(separator)s')"
+        sort_results = self.extra.get('sort_results', False)
+        if sort_results:
+            return "%(function)s(" \
+                   "array_agg(%(field)s), " \
+                   "'%(separator)s'" \
+                   "ORDER BY %(field)s" \
+                   ")"
+        else:
+            return "%(function)s(array_agg(%(field)s), '%(separator)s')"
 
 
 PostgreSQLGroupConcat = PostgreSQL9GroupConcat